Reorganize the global server state's social.rs into itself, parties and linkshells
